# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ent.Diagnose:
    properties:
      edges:
        $ref: '#/definitions/ent.DiagnoseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DiagnoseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.DiagnoseEdges:
    properties:
      doctor:
        $ref: '#/definitions/ent.Doctor'
        description: Doctor holds the value of the doctor edge.
        type: object
      illness:
        $ref: '#/definitions/ent.Illness'
        description: Illness holds the value of the illness edge.
        type: object
      patient:
        $ref: '#/definitions/ent.Patient'
        description: Patient holds the value of the patient edge.
        type: object
      symptoms:
        $ref: '#/definitions/ent.Symptoms'
        description: Symptoms holds the value of the symptoms edge.
        type: object
    type: object
  ent.Doctor:
    properties:
      Doctor_Email:
        description: DoctorEmail holds the value of the "Doctor_Email" field.
        type: string
      Doctor_NAME:
        description: DoctorNAME holds the value of the "Doctor_NAME" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DoctorEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DoctorQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.DoctorEdges:
    properties:
      doctorDiagnose:
        description: DoctorDiagnose holds the value of the doctor_diagnose edge.
        items:
          $ref: '#/definitions/ent.Diagnose'
        type: array
    type: object
  ent.Illness:
    properties:
      Illness_Name:
        description: IllnessName holds the value of the "Illness_Name" field.
        type: string
      edges:
        $ref: '#/definitions/ent.IllnessEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the IllnessQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.IllnessEdges:
    properties:
      illnessDiagnose:
        description: IllnessDiagnose holds the value of the illness_diagnose edge.
        items:
          $ref: '#/definitions/ent.Diagnose'
        type: array
    type: object
  ent.Patient:
    properties:
      Patient_Name:
        description: PatientName holds the value of the "Patient_Name" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PatientEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PatientQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PatientEdges:
    properties:
      patientDiagnose:
        description: PatientDiagnose holds the value of the patient_diagnose edge.
        items:
          $ref: '#/definitions/ent.Diagnose'
        type: array
    type: object
  ent.Symptoms:
    properties:
      Manner:
        description: Manner holds the value of the "Manner" field.
        type: string
      edges:
        $ref: '#/definitions/ent.SymptomsEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the SymptomsQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.SymptomsEdges:
    properties:
      symptomsDiagnose:
        description: SymptomsDiagnose holds the value of the symptoms_diagnose edge.
        items:
          $ref: '#/definitions/ent.Diagnose'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SA Example API Playlist Vidoe
  version: "1.0"
paths:
...
  /symptomss:
    get:
      description: list symptoms entities
      operationId: list-symptoms
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Symptoms'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List symptoms entities
    post:
      consumes:
      - application/json
      description: Create symptoms
      operationId: create-symptoms
      parameters:
      - description: Symptoms entity
        in: body
        name: symptoms
        required: true
        schema:
          $ref: '#/definitions/ent.Symptoms'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Symptoms'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create symptoms
  /symptomss/{id}:
    delete:
      description: get symptoms by ID
      operationId: delete-symptoms
      parameters:
      - description: Symptoms ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a symptoms entity by ID
    get:
      description: get symptoms by ID
      operationId: get-symptoms
      parameters:
      - description: Symptoms ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Symptoms'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a symptoms entity by ID
    put:
      consumes:
      - application/json
      description: update symptoms by ID
      operationId: update-symptoms
      parameters:
      - description: Symptoms ID
        in: path
        name: id
        required: true
        type: integer
      - description: Symptoms entity
        in: body
        name: symptoms
        required: true
        schema:
          $ref: '#/definitions/ent.Symptoms'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Symptoms'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a symptoms entity by ID
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"